rip out more types and replace some usage in rpc-client and rpc-client-api
