+++ account-decoder/src/parse_token.rs
@@ -1,8 +1,11 @@
+pub use solana_account_decoder_client_types::{
+    real_number_string, real_number_string_trimmed, UiTokenAmount,
+};
 use {
     crate::{
         parse_account_data::{ParsableAccount, ParseAccountError, SplTokenAdditionalData},
         parse_token_extension::{parse_extension, UiExtension},
-        StringAmount, StringDecimals,
+        StringAmount,
     },
     solana_sdk::pubkey::Pubkey,
     spl_token_2022::{
@@ -171,57 +174,6 @@ impl From<AccountState> for UiAccountState {
             AccountState::Uninitialized => UiAccountState::Uninitialized,
             AccountState::Initialized => UiAccountState::Initialized,
             AccountState::Frozen => UiAccountState::Frozen,
-        }
-    }
-}
-
-pub fn real_number_string(amount: u64, decimals: u8) -> StringDecimals {
-    let decimals = decimals as usize;
-    if decimals > 0 {
-        // Left-pad zeros to decimals + 1, so we at least have an integer zero
-        let mut s = format!("{:01$}", amount, decimals + 1);
-        // Add the decimal point (Sorry, "," locales!)
-        s.insert(s.len() - decimals, '.');
-        s
-    } else {
-        amount.to_string()
-    }
-}
-
-pub fn real_number_string_trimmed(amount: u64, decimals: u8) -> StringDecimals {
-    let mut s = real_number_string(amount, decimals);
-    if decimals > 0 {
-        let zeros_trimmed = s.trim_end_matches('0');
-        s = zeros_trimmed.trim_end_matches('.').to_string();
-    }
-    s
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct UiTokenAmount {
-    pub ui_amount: Option<f64>,
-    pub decimals: u8,
-    pub amount: StringAmount,
-    pub ui_amount_string: StringDecimals,
-}
-
-impl UiTokenAmount {
-    pub fn real_number_string(&self) -> String {
-        real_number_string(
-            u64::from_str(&self.amount).unwrap_or_default(),
-            self.decimals,
-        )
-    }
-
-    pub fn real_number_string_trimmed(&self) -> String {
-        if !self.ui_amount_string.is_empty() {
-            self.ui_amount_string.clone()
-        } else {
-            real_number_string_trimmed(
-                u64::from_str(&self.amount).unwrap_or_default(),
-                self.decimals,
-            )
         }
     }
 }
