+++ rpc-client/src/mock_sender.rs
@@ -5,7 +5,8 @@ use {
     async_trait::async_trait,
     base64::{prelude::BASE64_STANDARD, Engine},
     serde_json::{json, Number, Value},
-    solana_account_decoder::{encode_ui_account, UiAccountEncoding},
+    solana_account_decoder::encode_ui_account,
+    solana_account_decoder_client_types::UiAccountEncoding,
     solana_rpc_client_api::{
         client_error::Result,
         config::RpcBlockProductionConfig,
